Relocate mergeByField
